backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
PyJWT[crypto]==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
redis==5.2.1
python-multipart==0.0.20
httpx==0.28.1
